PR 2.7: bands collaborator migration (last Stage 2 PR)
Move BandMetadata class (~1075 LoC) from
src/pyramids/dataset/ops/band_metadata.py into
src/pyramids/dataset/_collaborators.py:Bands. Migrated members:

* Public methods: get_attribute_table, set_attribute_table, add_band,
  get_band_by_color, change_no_data_value.
* Public properties: band_color, color_table (both read/write).
* Private helpers: _attribute_table_to_df, _calculate_bbox,
  _calculate_bounds, _change_no_data_value_attr, _check_no_data_value,
  _df_to_attribute_table, _get_band_names, _get_color_table, _iloc,
  _set_band_names, _set_color_table, _set_no_data_value,
  _set_no_data_value_backend.

Substituted self.<state> -> self._ds.<state> for: _band_names,
_no_data_value, _raster, _update_inplace, access, band_count,
cell_size, columns, dtype, epsg, gdal_dtype, logger, no_data_value,
numpy_dtype, raster, read_array, rows, top_left_corner.

Five fixups required by the migration:

* change_no_data_value(inplace=True) used to ``return self``; now
  returns None and the Dataset.change_no_data_value facade substitutes
  ``self`` for None (matches apply / fill).
* add_band's classmethod-access ``type(self)._df_to_attribute_table``
  uses ``Bands._df_to_attribute_table`` directly (it's a staticmethod).
* The Dataset band_names setter and no_data_value setter route through
  self.bands._set_band_names / self.bands._change_no_data_value_attr.
* NetCDF.no_data_value setter delegated to ``self._change_no_data_value_attr``;
  now uses ``self.bands._change_no_data_value_attr``.
* AbstractDataset.__init__ calls _get_band_names / _calculate_bbox /
  _calculate_bounds before the Bands collaborator is wired. Added
  concrete Dataset overrides (mirroring the Bands implementations) so
  these work during initialization. Added a top-level
  ``import geopandas as gpd`` for _calculate_bounds.

src/pyramids/dataset/dataset.py:

* Drop _BandMetadataMixin from the import alias list and from Dataset's
  MRO. With this PR the MRO is just ``Dataset(AbstractDataset)`` —
  every public op now lives on a collaborator, and Dataset is a thin
  facade.
* Add five method facades plus band_color / color_table read/write
  facade properties.
* Add concrete overrides for the abstract methods _check_no_data_value,
  _set_no_data_value, _read_block (via collaborator), and the
  init-time methods _get_band_names, _calculate_bbox, _calculate_bounds
  (direct on Dataset).

src/pyramids/dataset/ops/__init__.py: __all__ is now empty — no mixin
classes remain in the public surface. ops/band_metadata.py: deleted.

tests/:

* test_dataset_unit.py / test_dataset.py: tests calling private Bands
  helpers (_set_no_data_value_backend, _change_no_data_value_attr,
  _set_no_data_value, _check_no_data_value, _set_band_names,
  _df_to_attribute_table, _attribute_table_to_df, _set_color_table,
  _get_color_table) route through ds.bands.<X>; class-level
  Dataset.bands._<X> sites use Bands.<X> directly. patch.object
  targets switched from ds to ds.bands.
* test_collaborators.py: FORWARDING_METHODS is now empty (every public
  collaborator method has migrated), bands rows are in FACADE_METHODS,
  band_color/color_table joined FACADE_PROPERTIES.

Full main test suite passes; only the three pre-existing Windows
file-lock flakes in test_dataset_collection.py remain failing.
