+++ src/pyramids/netcdf/netcdf.py
@@ -410,9 +410,9 @@ class NetCDF(Dataset):
         """
         if isinstance(value, list):
             for i, val in enumerate(value):
-                self._change_no_data_value_attr(i, val)
+                self.bands._change_no_data_value_attr(i, val)
         else:
-            self._change_no_data_value_attr(0, value)
+            self.bands._change_no_data_value_attr(0, value)
 
     @property
     def file_name(self):
