+++ tests/dataset/test_collaborators.py
@@ -233,18 +233,17 @@ class TestCollaboratorAttachment:
 # Dataset method (the mixin is still in Dataset's MRO). PR 2.1 (cell)
 # migrated method bodies into the collaborator and inverted the
 # direction — those methods are tested in TestFacadeDelegation below.
-FORWARDING_METHODS = [
-    ("bands", "get_attribute_table"),
-    ("bands", "set_attribute_table"),
-    ("bands", "add_band"),
-    ("bands", "get_band_by_color"),
-    ("bands", "change_no_data_value"),
-]
+FORWARDING_METHODS: list[tuple[str, str]] = []
 
 # Stage 2 facades: Dataset method delegates to the collaborator method
 # (the mixin has been removed from Dataset's MRO). PR 2.1 — cell, PR 2.2 —
 # cog, PR 2.3 — vectorize, PR 2.4 — analysis, PR 2.5 — spatial.
 FACADE_METHODS = [
+    ("bands", "get_attribute_table"),
+    ("bands", "set_attribute_table"),
+    ("bands", "add_band"),
+    ("bands", "get_band_by_color"),
+    ("bands", "change_no_data_value"),
     ("io", "read_array"),
     ("io", "write_array"),
     ("io", "to_file"),
@@ -359,16 +358,15 @@ class TestFacadeDelegation:
 
 READONLY_PROPERTIES = []
 
-READWRITE_PROPERTIES = [
-    ("bands", "band_color"),
-    ("bands", "color_table"),
-]
+READWRITE_PROPERTIES: list[tuple[str, str]] = []
 
 # Stage 2 facade properties: Dataset property delegates to a same-named
 # property on the collaborator. PR 2.2 — cog.is_cog. PR 2.6 — io.overview_count.
 FACADE_PROPERTIES = [
     ("cog", "is_cog"),
     ("io", "overview_count"),
+    ("bands", "band_color"),
+    ("bands", "color_table"),
 ]
